Combine SPDM app and IMAGE loader App to one (#221)
- Combine SPDM app and IMAGE loader App to one into 1 user app
- The user app uses 1 executor that spawns the spdm and image loader
tasks
