+++ Cargo.toml
@@ -26,9 +26,8 @@ members = [
     "platforms/emulator/runtime",
     "platforms/emulator/kernel/capsules",
     "platforms/emulator/kernel/drivers/dma",
-    "platforms/emulator/runtime/userspace/apps/image-loader",
     "platforms/emulator/runtime/userspace/apps/example",
-    "platforms/emulator/runtime/userspace/apps/spdm",
+    "platforms/emulator/runtime/userspace/apps/user",
     "platforms/fpga/config",
     "platforms/fpga/rom",
     "platforms/fpga/runtime",
