+++ platforms/emulator/runtime/userspace/apps/user/src/image_loader/pldm_fdops_mock.rs
