+++ platforms/emulator/runtime/userspace/apps/user/src/spdm/config.rs
