+++ platforms/emulator/runtime/userspace/apps/user/src/spdm/mod.rs
@@ -1,18 +1,12 @@
 // Licensed under the Apache-2.0 license
 
-#![cfg_attr(target_arch = "riscv32", no_std)]
-#![cfg_attr(target_arch = "riscv32", no_main)]
-#![feature(impl_trait_in_assoc_type)]
-#![allow(static_mut_refs)]
-
 mod config;
 mod dev_cert_store;
-
 use core::fmt::Write;
 use dev_cert_store::{DeviceCertChain, DeviceCertStore};
 use libsyscall_caliptra::mctp::driver_num;
+use libsyscall_caliptra::DefaultSyscalls;
 use libtock_console::{Console, ConsoleWriter};
-use libtock_platform::Syscalls;
 use spdm_lib::codec::MessageBuf;
 use spdm_lib::context::SpdmContext;
 use spdm_lib::protocol::*;
@@ -31,52 +25,21 @@ static HASH_PRIORITY_TABLE: &[BaseHashAlgoType] = &[
     BaseHashAlgoType::TpmAlgSha256,
 ];
 
-#[cfg(target_arch = "riscv32")]
-mod riscv;
-
-#[cfg(not(target_arch = "riscv32"))]
-pub(crate) fn kernel() -> libtock_unittest::fake::Kernel {
-    use libtock_unittest::fake;
-    let kernel = fake::Kernel::new();
-    let console = fake::Console::new();
-    kernel.add_driver(&console);
-    kernel
-}
-
-#[cfg(not(target_arch = "riscv32"))]
-fn main() {
-    // build a fake kernel so that the app will at least start without Tock
-    let _kernel = kernel();
-    // call the main function
-    libtockasync::start_async(start());
-}
-
-#[cfg(target_arch = "riscv32")]
 #[embassy_executor::task]
-async fn start() {
-    async_main::<libtock_runtime::TockSyscalls>().await;
-}
-
-#[cfg(not(target_arch = "riscv32"))]
-#[embassy_executor::task]
-async fn start() {
-    async_main::<libtock_unittest::fake::Syscalls>().await;
-}
-
-pub(crate) async fn async_main<S: Syscalls>() {
-    let mut console_writer = Console::<S>::writer();
+pub(crate) async fn spdm_task() {
+    let mut console_writer = Console::<DefaultSyscalls>::writer();
     writeln!(console_writer, "SPDM_APP: Hello SPDM async world!").unwrap();
 
     writeln!(console_writer, "SPDM_APP: Running SPDM-APP...").unwrap();
 
     let mut raw_buffer = [0; MAX_MCTP_SPDM_MSG_SIZE];
 
-    spdm_loop::<S>(&mut raw_buffer, &mut console_writer).await;
+    spdm_loop(&mut raw_buffer, &mut console_writer).await;
 
     writeln!(console_writer, "SPDM_APP: app finished").unwrap();
 }
 
-async fn spdm_loop<S: Syscalls>(raw_buffer: &mut [u8], cw: &mut ConsoleWriter<S>) {
+async fn spdm_loop(raw_buffer: &mut [u8], cw: &mut ConsoleWriter<DefaultSyscalls>) {
     let mut mctp_spdm_transport: MctpTransport = MctpTransport::new(driver_num::MCTP_SPDM);
 
     let max_mctp_spdm_msg_size =
